odoo-mcp-server/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
httpx==0.25.0
aiofiles==23.2.1
python-multipart==0.0.6
websockets==11.0.3
asyncio-mqtt==0.11.1
xmlrpc.client
requests==2.31.0
python-dotenv==1.0.0